refactor: box선택 시 클릭 가능하도록 로직 이동
