+++ client/src/components/FileUpLoader/FileUpLoaderBox.tsx
@@ -3,22 +3,33 @@ import { s_fileUpLoaderBox } from './FileUpLoader.style';
 import { useFileUpLoaderContext } from './FileUpLoaderContext';
 import { FileUpLoaderEmptyView } from './FileUpLoaderEmptyView';
 import { FileUpLoaderItemView } from './FileUpLoaderItemView';
+import type { KeyboardEvent } from 'react';
 
 function FileUpLoaderBox() {
-    const { files, hasFile, isActive, handleDragStart, handleDragEnd, handleDrop } =
+    const { files, hasFile, handleDragStart, handleDragEnd, handleDrop, handleClickButton } =
         useFileUpLoaderContext();
 
     const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
         e.preventDefault();
     };
 
+    const handleKeyDown = (event: KeyboardEvent<HTMLDivElement>) => {
+        if (event.key === 'Enter' || event.key === ' ') {
+            handleClickButton();
+        }
+    };
+
     return (
         <div
-            css={s_fileUpLoaderBox(hasFile, isActive)}
+            role="button"
+            css={s_fileUpLoaderBox(hasFile)}
             onDragEnter={handleDragStart}
             onDragLeave={handleDragEnd}
             onDragOver={handleDragOver}
             onDrop={handleDrop}
+            onKeyDown={handleKeyDown}
+            onClick={handleClickButton}
+            tabIndex={0}
         >
             {files?.length === 0 ? <FileUpLoaderEmptyView /> : <FileUpLoaderItemView />}
         </div>
